chore: refactor more for readability
